backend/requirements.txt
flask
flask-cors            # Allow frontend (React) to talk to backend
requests              # API calls
Pillow                # Image handling
reportlab             # Export stories to PDF
python-dotenv         # Manage API keys
cohere                # For Cohere API
os                     # Environment variable management
base64                # Image encoding
